skills/resend/pyproject.toml
[project]
name = "resend-email"
version = "0.1.0"
description = "CLI for sending emails via Resend"
requires-python = ">=3.11"
dependencies = [
    "typer",
    "resend",
    "python-dotenv",
    "hayeah",
]

[tool.uv.sources]
hayeah = { path = "../../hayeah", editable = true }

[project.scripts]
resend-email = "resend_email.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/resend_email"]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
    "pyright",
]

[tool.pytest.ini_options]
testpaths = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pyright]
pythonVersion = "3.11"
include = ["src"]
typeCheckingMode = "basic"
